requirements.txt
streamlit==1.42.2
torch==2.6.0
diffusers==0.32.2
numpy==2.2.3
requests==2.32.3
python-dotenv==1.0.1
deep-translator==1.11.4
psycopg==3.2.5
psycopg-binary==3.2.5
SQLAlchemy==2.0.38
pydantic==2.10.6
pydantic-settings==2.8.0
fsspec==2025.2.0
Pillow==11.1.0
rich==13.9.4
protobuf==5.29.3
pypdf
duckduckgo-search
arxiv
ollama
